requirements.txt
streamlit
pandas
numpy
pyswisseph
openpyxl
qrcode
Pillow
